fix placeholder input fields cannot be empty
The input fields in the placeholder return a string,
which is converted to a number using `parseInt`.
When the input field is empty, an empty string `""` is returned.
`parseInt` converted `""` to number, got `NaN`, used fallback: "2".
Which meant the input fields cannot be empty.
That was fixed by checking the string when the value changed
and passing `""` or the parsed number as needed.
Some types were also changed to accomodate this change.
I was pleasantly surprised to find out typescript is intelligent enough
to figure out the types of `initialRowCount` and `initialColCount`
from the `if` statement on line 49.