button el: add hover color support
This one has been a bit tricky.
Hover colors cannot be set using inline styles
and CSS files are not easy to change from JS.
Therefore, I reused a trick from wp-table-builder:
in CSS (`style.scss`), set a hover style that uses a CSS variable,
in the inline style, update the CSS variable to change hover styling.
Voila!
But an empty CSS variable simply reverts to white and
the button vanishes on hover; whereas, it should seemingly not change.
So, for example, if hover color for text is not set,
the text color is used. If the text color, in turn, is not set,
then the default color is used.
So, an `useEffect` hook is used to determine the value of the
hover CSS vars as the button background or text colors are changed.

This was my first time using `useEffect` in prod and traditionally,
I managed to get myself stuck on an infinite `useEffect` loop -
debugging which consumed more than an hour of my life.