cell: update content align support to use floats
`text-align` does not work on all elements. Namely,
button element, which has a `inline-block` inside a block.
We could go for the modern flexbox/grid approach but then,
the whole table would need to be emulated with flex/grid.
Because flexbox and grid don't play nice with older table layouts.