button el: changed alignment from float to margin
Using `float: left` or `float: right` makes multiple buttons
stay side by side, whereas, the expected behaviour is for
them to stack on top of each other. To achieve that, floats
have been dropped in favour of `margin: auto`. To align a
button to the right, its `margin-left`has to be set to `auto`
and vice-versa for left alignment. To center a button, both
the `margin-left` and `margin-right` has to be set to `auto`
- shorthand for which is `margin-inline: auto`.