cell: rename css classes from justify to align
The justify controls were inspired from the `core/button`
block, which uses flexbox for alignment/justification.
Therefore, I initially named the classes `justify-...`. Since
we're not using flexbox for alignment, and we're not using
the word "justify" for alignment anywhere in the UI, we're
going to rename the classes to use "align" for avoiding
confusion.