button el: add features - URL, URL rel, URL target
These three features are stored in three attributes `url`,
`linkTarget`, and `rel`.
Most of the code is straight up copied over from `core/button`.
We'll have to keep an eye at `core/button`'s use of `useMemo`
for resolving a conflict with LinkControl as it is marked as
subject to change.