fix issue with some older tables not rendering
If an attribute has a default value, it is not saved in the post data.
So, if an attribute's default value is changed in block.json, all blocks
with the previous default value will now have the new default value.
This is what happened while removing the `hasTableCreated` attribute.
This commit changes the code to rely on cells instead of rows and cols.